Improving standard compiler compatibility (#2926)
* Adding explicit dolphin/ prefix & fix characters

* Rename ShiftJIS to SJIS

* Separate JASSeqReader read methods implementation between compilers.

* Fix pointer.h

* fix d_item_data typo

* fix gcn matching issue

+++ include/JSystem/JStudio/JStudio/fvb-data-parse.h
@@ -27,7 +27,7 @@ public:
         return ret;
     }
     const void* getContent() const {
-        return (const void*)((int)getBlockEnd_() + align_roundUp(get_IDSize(), 4));
+        return (const void*)((intptr_t)getBlockEnd_() + align_roundUp(get_IDSize(), 4));
     }
 };
